Add in SoundController system with Layered Sound capability
Added in SoundController.cs and Sound.sc, a very flexible and easy to use system for playing sounds in the game.

Original task was implimenting Layered Sound which is now done using this controller, but I would also like to use this to replace other implimentations of sound effect playing in the game.