Removed Transparency GameObject from all scenes and added details rule tile
There was a GameObject called transparency in every scene that made the background look less colorful. They have all been deleted except for in Chapter1 Level1. The transparency GameObject in Chapter1 Level1 has been disabled, but it does make the background contrast better with the player so it can still be enabled if we deem that it looks better with it.

- Added Lab Details Rule Tile for adding details to the tilemap. The proper use of this rule tile is to create a second tilemap as a child of the Grid GameObject, then draw the details on this detail tilemap rather than the base tilemap.