Add files via upload
The new design for Level 3 from the external build is now technically in the game, but still requires some work in order to make it fully function. 

(1) Backgrounds were not transferred properly. 
(2) Moving Platforms and Batteries do not function; recommend porting those as well, but that may screw things up in other levels, so don't do it without asking anyone.
(3) Enemies do not function. 
(4) Spike Pits use a prefab that does not exist outside of the External Build, transfer that ASAP
(5) Level exit needs to be reattached to Level 4.