Switch to URP / Lighting Demo Scene
Implementation of a switch to URP and addition of 2d lighting in the lighting demo scene. Some work on implementing glow effect for the Atlas and the background coral. Glow is only on the default chatacter sprite and not the animations though. Also implemented lighting and partical effect on an outlet, the idea being that that kind of a thing could be applied as a diagetic way to show what outlet was focused.