GS-1112 Implementing wire despawn
Implemented a player respawn event in LevelManager to handle when the player needs to respawn at a checkpoint. Added wire despawning to fix the bug where the wire was still spawned after resetting to a checkpoint.

Ideally, we'd also have the actual teleportation handled in the MovementExecutor as well - we can do this by having a proper checkpoint system and getting the most recent checkpoint, or by passing in the respawnPoint position through the UnityEvent.