Fixed GameCamera Bug in Template Scene
When the Pixel Perfect Camera Component  of the Main Camera of the GameCamera Prefab was updated to the URP Pixel Perfect Camera Component, the Universal Additional Camera Data Script was duplicated, causing a bug that would make the scene appear blank. The GameCamera Prefab needed to be deleted and readded to the scene.