Changed TrueParallax cam to use Camera.Main.gameobject instead of serializing a reference
The TrueParallax Script was using a [SerializeField] private GameObject cam, which required the user to drag the camera into the field on the script. When you have a lot of objects with this TrueParllax Script, it becomes tedious, annoying, and a waste of time to assign the camera to each of these scripts. It now simply get's the GameObject of the Camera.Main