Changed sound system to not depend on SoundController
Previously, many Prefabs depended on the existence of a SoundController in the scene, and required the SoundController to be assigned to the scripts that required it. This becomes tedious as you add more things to a scene. Furthermore, the SoundController script is hard coded to reference all of the sound files it needs by getting the file name of the scriptable objects that contains the sound files that it needs from the Resources folder. This is very unscalable and would likely cause problems as more sounds are added to the game. This new implementation (which should be reviewed) Has a new type of ScriptableObject called SoundArray. A SoundArray holds an array of Sound ScriptableObjects. The SoundArray ScriptableObject has a public method that looks through the array for a Sound ScriptableObject with the field soundName that matches the given string. This removes scene dependencies, and is more scalable than the SoundController.