GS-1041 Fix collider on NPC Robot
There was a bug caused the player to get stuck on/under the enemy robot. The RobotSpriteRenderer (child object of NPCRobot prefab) has a capsule collider on it that's a trigger, but is still being detected by CollisionDetector scripts on the player (because the boolean collidesWithTriggers is being ignored in that script).

This fixes the bug by creating a new layer called NPCs that the player's collision detectors don't track, but we should eventually solve the underlying issues of the CollsiionDectector script.

Also, not sure if we still need the capsule collider on RobotSpriteRenderer at all... can likely be removed