Removed use of missing shader and Added NormalMap Tileset Test to LightingTest Scene
There was a shader for a glowmap that didn't get preserved when the lighting test scene was merged. I removed it since it never worked correctly anyway.
Created a normal map for the metal tileset and added it as a secondary texture. Then turned on normal maps for the lights in the test scent. To see it look at the platform in the first pit. There is a problem with the way that the normal map is cut up for the tileset that doesn't work quite right. Not sure if I like the effect, but that's why its a test scene :)