Wide edge bubble particle emitters for both layers of bubbles
Added edge particle emitters the full width of the level. The other emitters stack well with these. The new ones are the front item on each layer of the paralax bg which looks pretty good for bubbles outside a window. Maybe they would look better as the back layer for rain.