Created DiegeticOutletPlacementTest Scene
Created a scene showing what could be a diegetic explanation for outlet positioning

By placing outlets ON TOP OF the ground tilemap instead of just randomly in the air, there is a more easily explainable diegetic reasoning for the outlet existence in the world. This however would require the wire to be able to reach the outlets through colliders. I think this would be a worthwhile change however, because currently when the wire hits a collider, it just gets retracted, which feels significantly worse than connecting to the outlet anyways. I also created a temporary alternative tp the current platform we have, where the outlet is located on the platform, rather than on a terminal, allowing the platform to be controlled while the player is on it.