Collision detection updates (#9)
* testing out different collision detection

* Updating player prefab for collisions

Fixing up some of the collision detection by altering how we are searching for colliders on layers, and actually ignoring triggers. Note that because this now takes the layer mask into account correctly, there will need to be some fixes for other prefabs (i.e. outlets, doors, etc.) to make sure they are on the correct layer.