GS-1085 Fixing no movement on platform bug
I think this should fix the stationary Atlas getting left behind on the platform bug... I increased the size of the platform collider slightly, and also added in a tag check for the platform script.

The way the platform collision works should really be handled entirely on the player side, but it's just a little more complex because of the way the collision detectors work... that should get fixed if we have time.