Updated Battery Pack Prefab
Set the Max Charge and Max Virus fields of the battery pack prefab to make it so that when you drag in the prefab, you don't have to set them yourself. They were at zero before, so even though it was holding 50 energy, it's max was set at 0 so it wasn't working. Now it's set to start with 50 energy with a max of 50 energy