Adjusting fall to wire swing time
Player would sometimes annoyingly walk off the edge while connected to the wire, and the wire would be too long so the player would get end up swinging into the wall and wouldn't be able to get back up to the ledge. adjusted the fall to wire swing time to be less.