Updated Tileset Again
Updated the metal ground tileset art again to make the tiles seamless. Not all of the tiles are seamless, but most of the ones used when making levels are seamless. Most seaming issues can be fixed by making the tiles more of a "blob". Some tiles may need to be adjusted in the future.