Added Dialogue and Portrait for Scrap of Friend
The scrap, when interacted with, now has its own dialogue describing the scrap of the friend and relating that Atlas remembers them. This implements a new portrait, showing, in more detail, the scrap of your old friend.
In the future, this could be changed to a different styling of on-screen graphic.