Committed the Team 5 submission for the GMTK puzzle exercise
Implemented a basic puzzle as a group which combines battery draining deactivating devices and moving platforms. After hitting the play button, hit the pause button to start the level. The solution is to drain each moving platform at the right time to make a series of jumps the player can make to the right side of the level.