Created Level 3A Prototype
Created a new scene, Level3APrototype, in the Level Ideas and Exploration subfolder of the Scenes folder in the Assets folder.
This level roughly prototypes the spatial organization and platforming progression of a player through the proposed Level 3A revision.
The design of this level was based on the Level 3A Overview and Layout 2023 Fall, which can be found in the Room 3 Safe Area folder of the Design By Chapter folder of the Design folder.