Implemented all functional values for level elements, began aesthetic details
Over the course of this sprint, all functional values for all level elements had their parameters set to correspond with the design for Level 3A proposed in previous sprints. This means that, in theory, if there are no bugs in implementation of the game's core mechanics, the level would be fully functional.
Additionally, some aesthetic refining has been begun on the level in this commit, though it is not completed yet.