Created easter egg in room2a.
I created a new variation of the level with the easter egg, located at the first vertical elevator. The easter egg is supposed to hint at the overarching narrative. (In the future, I hope to implement the dialogue for the dead robot) The level is in the level ideas folder.