Started Art Department's Unity Folder System
Matched, to the best of my ability, the files already renamed in the Master Art Asset google sheet to those that exist in the Unity project. Renamed those files and moved them to their temporary home in Visuals/LIVE/GLOBAL (new folders). Will check over them and resort if needed.