Updated WarningAnimation script to work with custom animations
Changed WarningAnimation.cs so that it plays the default blinking animation if the warning object doesn't contain an Animator component. If it does contain an Animator component, then it enables the Animator instead, playing a custom animation. Also created Warning.controller and WarningAnimation.anim as a sample controller and animation.