Revised core energy/virus system through AControllable to better conform to documented design
Notably, this revision only affects the AControllable script and does not affect any other part of the codebase directly, thereby requiring no refactoring.
The change itself is just ensuring that the rate of energy/virus transfer is proportional to the virus proportion present in the source that is being drawn from. This has the added benefit of also increasing the precision of our float values tracking energy and virus. Examples of the new system being tested and confirmed working can be scene in the EnergyVirusResearch scene modified by this commit.