patched URP Blit warnings from art asset refactor
Due to an art asset "Virus Hit Overlay" material being renamed or otherwise dereferenced, the URP Blit scriptable object lost a field it needed to function. On GUI updates, the editor console would be filled with warnings. To fix this, I searched up the asset by GUID (using https://forum.unity.com/threads/how-to-find-object-by-guid-in-project-window.417608/) and replaced it. It's called "Virus Hit Overlay 1," which to me indicates that it was duplicated but not reassigned.