Moved elements in 2a and Updated WarningAnimation
Moved the background elements in scene 2a to fit with the narrative changes, as well as to make the elements more visible and realistic. Also, made minor changes to WarningAnimation so that the blinking animation works with the SpriteRenderer instead of UI.Image, as the virus warning will be a GameObject with a SpriteRenderer.