Updated Controllable structure to add StallingDoor. Implemented StallingDoor in Level3APrototype.
Modified AControllable and ControllableDoor scripts to be as accurate as possible to AControllable's new interpretation of energy and virus from the PlayerInfo SO. As an experiment, created StallingControllableDoor as an example of how existing level element code needs to be changed to implement virus response functionality. Integrated this functionality into Level3APrototype to help bring the scene closer to the intended documented design goals.