Added Virus Eye and Virus Headlight warnings
Edited player animations to include keyframes for the Virus Eye asset, which would be a child of PlayerGraphics (this doesn't affect any animations for when Virus Eye is not a child of PlayerGraphics). Changed WarningAnimations and PlayerHealth scripts so that as the player gains more virus, their headlight and eyes become more purple, and at high percentages, the headlight starts blinking.