Created soft-reset feature, activated with "[" button press
By hitting this button in any level, the entire level will be reset back to its beginning state, but the player will spawn in at the most recent checkpoint in that level that they hit. In this way, soft lock scenarios can be reset out of without the player losing the progress they made in the level.