Finished basic visual revision and functional implementation of Level 3A
Level 3A has had a base visual overhaul to make it more consistent with the original level it was based on.
Level 3A is technically fully functional and can be loaded into from the previous scene and can be traversed to load into the next scene.
However, the unique features in the level definitely still need some more testing to refine, and we need to do blind tests to check how effective the level is as an onboarding tool.