Updating Camera Prefab to Use Camera Lookahead
Enabled camera lookahead with smoothing and damping values to mostly lookahead on the x axis but also on the y axis for falling and elevators. I think I got the damping to the point where swinging on the grapple is ok. I still think the reset is a little jarring but the deadzone is such that it's only a little movement...