Updated Clear Elevator Pipe
Made a more complex game object from the pipe sprite so tat atlas will be inside of the pipe above the doorway. Added a elevator direction arrow and a detail to the top of the pipe. Need to add a triggered platform inside that freezes player movement and moves up. Need to move the level transition trigger to be above the elevator.