Replaced Elevator in level 2a with the cu
The level transition in level 2a was turned into an elevator that needed power - it is now a trigger which starts a cutscene once Atlas gets into the elevator. This new exit elevator is now a prefab that can be placed into future scenes. You still have to drag the player and cameras into the animation trigger,

Right now, 2a is the only level with an elevator at the end. If the final plan is to have all the levels end with an elevator, this prefab should probably be changed to support different sprites depending on the background.