Boss Fight Level Visual Update
Replaced the old door with the new elevator. Added a waves icon to the elevator (surface). Moved turrets around, adjusted the camera (set to look right on full hd or 4k). Lowered the amount of health in the turrets to make turning them off easier. It's still not "great" but it's a little more playable. The big thing that needs to be added is the voice module. I think moving the battery station to the top right and making the space below the screen closed with the voice module in it. Then removing its doors at the end of the fight. Then making the exit door only open if you have the voice module. Maybe making it a NPC with dialogue that says it needs the voice module, and then after say access granted and open/triger the leaving cut scene.