Added the voice command collectable to dictionsry
EventCollectables.cs is in the Depricated folder. That is probably because it was just not being used yet. Need to move it back as part of cleanup for this task. Now picking up the voice module in the center of the boss scene will grant the InkGlobal var voiceModuleObtained. Added a tmp NPC to the start of the level to test it. That works. NPC needs to be removed or to have different dialogue. Exit door needs to trigger off of the voiceModuleObtained. Adding the NPC also added the InkDialogue system to the game which needed to happen anyway. I may create a tmp. Jones NPC to add the trash talking that is in the documentation for the scene.