Working Boss Chamber Exit
The exit triggers a dialogue saying you need the voice module. Killing the boss opens the doors to the module but not the exit. Collecting the voice module updates the globals dictionary. Colliding with the door after getting the voice module opens it. If you stand in the door and interact you get the exit dialogue. It is not triggering automatically if you have hit it before in the force state. Need to find a way to make dialogue auto-trigger when there is a state change...