Hack to solve the exit dialogue problem
Added a second trigger just inside the door collider so that you can't hit it till you are opening the door. Works well, kind of a hack. Need to add dialogue for Jones and have it dependent on not having the voice module. Then it will stop when you take the module. Also need to deactivate the turrets when you overload Jones.