Retracts the plug on collision
Added a function the IPlug class that takes a boolean value to represent the retracting value of the plug. It is implemented in the straight class. It sets the retracting variable to the parameter of the function. The plug executioner script calls this function when there is a collision with a trigger or collider on the default layer.