plug hit box smaller + boss room camera change
Made a change to the plug prefab so that the collider is smaller. Was a nice QOL change so that the plug collisions match the visuals more. Also removed the collider on the camera in the boss room. Was stopping the plug from being fired now that it retracts when hitting colliders.