Josh GMTK Scene finally.
Finally got around to doing my GMTK scene. This should only be committing that scene but I've still been getting weird errors when running the game on my machine when interacting with the charging outlets where the game crashes and the error message says something like variable name Density not found that I hope doesn't spread to others from this. But again this commit should only contain the puzzle scene I worked on.