Initial Script for the rechargable and leaky batterys
This is the first iteration of this script. There are no compile errors but it has not been tested in game yet. It is a child class of an outlet. The additional function is a fixed update. If the player is connected this fixed update does nothing. If they're not connected the outlet will gain or lose charge based on if it's a leaky or charging battery.