Added effects to rechargeable and leaky outlets.
Added an array of 2d lights to the rechargeable outlet to play while recharging. Added an array of particle systems to play while the battery leaks. There is also a small bug fix in the charging outlet that checks for percentage full as .98 instead of 1 because the outlets don't actually fill up to 1