Initial implementation of the sparks on the wire.
This is currently not working. But I added a particle system to the player that will be used. Changed what the change energy event returns to what the change was not what the total is after the change. Added the controllable of the outlet as a variable to the wire thrower. Also added listeners to it. Added two functions for setting the particle system color and rendering the particle system.