Intercom asset 1.4.2 (Unfinished) + Laser Death Sound Correction
There should be no change to gameplay so far, but everything required for intercom visuals are now in the assets and there is an intercom in level 1.4.2. I'm pushing this now because I need to access a file that got left on an earlier branch. Additionally, I increased the base volume of the laser death sound so it can be heard as intended.