Address issue regarding downward moving platforms leading to stunting the player's velocity.
https://games-at-northeastern-studio.atlassian.net/browse/GS-5475
The bugfix works for all moving plugs(both moving element and moving outlet scripts). Moving outlets should be swapped out for moving element as desired in 1.5.2. The bugfix does not fully handle the issue for high moving downward platforms due to inconsistency in transferring the movement of the plug to the player. It works better than it did before under these cases, but it leads to the player being very jittery. If high speed platforms are wanted to be in the game, another bugfix would need to be made on top of this one, most likely finding out how to account for the difference in movement. Also, someone should most likely look into Kinematics.cs. Every use I found with it seems to lead to edge case issues.