Create Vox ambient dialogue for 1.5.2
Part of this task was to shorten Vox's dialogue at the very beginning of the level. As of now that intro dialogue is removed as the way it was implemented is not allowing new dialogue to take its place. But, no dialogue still works here as the ambient lines can begin whenever we want.
GS-5218
https://games-at-northeastern-studio.atlassian.net/browse/GS-5218?atlOrigin=eyJpIjoiYTA2YmVjODQzM2I5NDVkZmIxODQzYTEyYTQyYWQ1NmUiLCJwIjoiaiJ9