GS-5630 Implement Voice Module Door to 1.1.1
The Voice Module Door was added to the older 1.1.1 scene, it should be added to the updated 1.1.1 with the new tile map. It should cleanly be integrated with the wall

Deliverable: Add the voice module door, dialogue, and collider to 1.1.1 Training Facility 1. Also update the tile map to make it seem like it is “integrated” into the wall.