GS-6026
"Something was adjusted regarding the door script whereas if they have full power, the offset doesn’t occur when the scene is loaded, leading them to now stay in the ground. Alter their transforms to ensure they start blocking the player. " The door's initial movement based on power only occurs on change, not on initalization. This was adjusted as it seems like the new fix wouldn't conflict, but it may lead to some issues with 1.3 if those doors are adjusted aswell.