GS-6007
https://games-at-northeastern-studio.atlassian.net/browse/GS-6007

Added a door to the initial chamber that remains closed until the player speaks with the handler. Added methods to ControllableDoor to allow for dynamic "openDoor". User can call openDoor and now pass in a time parameter to open doors at custom timings. Also needed to bind this function in the Ink Dialogue Manager.