GS-6354: Fix Infinite Virus Laser Range on 1.5.1
https://games-at-northeastern-studio.atlassian.net/browse/GS-6354

The visual for and the actual damage from the range were called in two separate locations. I altered the code to ensure the line renderer logic is only activated if the player is in the maxdistance, tying the two together. This should affect the lasers in 1.5.2 as well if they are moved into longer sightlines.