GS-6481: Determine Discrepancies with Sprite Swapping for Outlets
https://games-at-northeastern-studio.atlassian.net/browse/GS-6481
Firstly, starting forcing some outlets to now be under the prefabs. Some change was made that reinterpreted the rates for change of energy for outletmeters that made old ones 10x slower than standard, so they were adjusted to become prefabs that have the proper energy in case it is changed again. Secondly, the controllables would allow for negative energy in some cases due to float arithmatic. Functions were added in to attempt to force it to be 0; however, it also may lead to issues in the future, so should continue to be looked at if any other bugs come out of it.