GS-6593: Update Wires on 1.5.2 to ensure their minimum energy matches with the objects they are controlling
https://games-at-northeastern-studio.atlassian.net/browse/GS-6593
I had to update the OutletWire code a bit to get it to what I want to do. Now it also has a toggle to use combined energy of the outlet, which is how a lot of secondaries in 1.5.2 work. They are done at 19.9 rather than 20 since floats would make them fail a fair amount of the time with their comparisons otherwise.