GS-6731: Design optional portions to 1.1 to expand on lore and exploration
https://games-at-northeastern-studio.atlassian.net/browse/GS-6731
Created a lower room as a corridor into the kind of underworkers of the level, emphasing the piping with lesso on the training aspect. The idea is to display more of the lore of the level rather than just of the generalized story(NPC left down there with failing navigation, a gangway with caution that is functional to contrast the one above, another exit with voice module to make the level feel like a larger facility, etc). The skill check is really just recognition of alternate routes and more training with basic swinging(the player just has to make a basic swing backwards rather than forwards), to hopefully get the more comfortable while still receiving small lore rewards.