GS-6054 Make Lights Flicker Level 3
Jira link:
 https://games-at-northeastern-studio.atlassian.net/browse/GS-6054

Attached a script to all of the lights in level 1.3, the server room, (Every light under rectangle windows and circle windows) that will make the lights flicker by being assigned random values of intensity between a minimum and maximum value. The script is only enabled for some of the lights, not all of them, so that some lights will not flicker.

The script can be found in Assets > Scripts > Level Specific Scripts > Chapter 1 > Level 3, and it is called FlickerControl. It should be very easy to modify.