GS-6367: Implement Dialogue Switching For Vox and Atlas Voice on the Reverse Progression
https://games-at-northeastern-studio.atlassian.net/browse/GS-6367

Implemented a basic UI and controller for the voice module dialogue switching. It should have no affect on the current gameplay, as all are deactivated until the last pickup in 1.5.2. Now that it is directly established in the player info rather than being set in level manager parameters, what was done as concept with the upgrade handler can now be ripped up and removed, but that would come in a seperate task in refactoring what is done in 1.5.2 regarding dialogue. A test scene should also be made for this within either the end of this sprint as QA, or next sprint to demonstrate its capabilties with an NPC.