GS-6368, GS-6371: Create a Test Scene That has an NPC for Atlas to Response to with the Voice Module Upgrade
https://games-at-northeastern-studio.atlassian.net/browse/GS-6368
https://games-at-northeastern-studio.atlassian.net/browse/GS-6371
These two tickets I did under one submission given one is an example over the other. This is not good practice, but to try to remove them from each other in this case would be time consuming for very little value. The system has been set up such that the voice module will work(with insert art waiting on the art tasks).