GS-7011 (Fix Buggy Outlet Extension in Specific Sections of 1.2)
The grapple currently collides with the elevators in the level creating a buggy effect if the player grapples near them. There was also a hidden game object in 1.2 that I deactivated. To fix the elevator bug I added a new layer that you can add to the elevators so it will ignore the grapple collision!