Redo Of: Remade Virus Laser Shader To Look Good While Not Being Blinding
This commit was redone because of merge conflicts and slight edits

I remade the shader for the virus laser so it retains its laser-like qualities, while not being overly blinding to the player. I also edited how the lasers were rendered, specifically with how the edges of a laser were rendered.