GS-7215: Unify moving outlet interaction with the new moving object script
https://games-at-northeastern-studio.atlassian.net/browse/GS-7215
Outlets for both moving scripts have been updated to use their actual velocity rather than a movement vector from the moving scripts, so they can generalize to whatever script is moving them. The swinging, now using that and fixedupdate properly, has also as a byproduct reduced the clankyness of movement on some sets of moving outlets, which is good. High speeds are still a bit inconsistent and the lack of a swing speed cap is really apparent now. External velocity has also been moved into a set of forces rather than a single force in preperation for dealing with swinging forces and to allow for updating a given force multiple times a frame.