GS-7101: Redo the bot and upgrades in the start of 1.5.2 to fall in line with the new voice module
Had to reinput some of the gameobjects onto vox to make sure the ending sequence was fine. There are still some bugs with dialogue cut early, but I believe those were here prior, so they will be dealt with in a seperate ticket. The upgrades have been replaced though and the dialogue is functioning as expected.