Fix Acceleration, Speed Calculations and Improved Platform Quality of Life Changes
Platform fixed to handle acceleration correctly and added the ability to toggle speed to 0 when the platform reaches either end of its path (PingPong Platform Only). Fixed the speed based on path length problem. Lastly, made it so the platforms subscribe to the path instead of the path holding references to each of the platforms