GS-7313 Polish and Implement New Puzzle Idea in Bottom Right of Level 1.4.1
Jira link: https://games-at-northeastern-studio.atlassian.net/browse/GS-7313

The goal of this task was to polish the puzzle concept that I came up with last sprint, and then implement this puzzle into level 1.4.1

This puzzle has replaced the puzzle that was in the bottom right corner of the level. It is a laser puzzle that involves using tracks to navigate through them.