Added ability for a moving object path to be a child of another path
The old moving object/platform implementation was able to have paths be children of other paths. While it hasn't been used in any level yet, I added this mechanic because I don't want mechanics to be lost. The only downside is that because of how physics calculations and transformations are calculated, objects with kinematic rigidbodies can't be children of another rigidbody and move with them. This would require a lot more work to recalculate how the rigidbodies interact. So, for now, the solution is putting the platforms side-by-side with the inner paths.