Fixed Performance of level 1.1.1
This is just a temporary change to 1.1.1 because of the performance errors. I figured out the problem: the lights in 2D are all calculated in real time. Because sprites don't have normals, you can't exactly bake lighting for a 2D scene. So, a high volume of lights all generating light per frame is what is tanking the framerate