GS-7491
Added Walls + Windows to the first quarter of 1.5.1. Don't seem to be any layering issues + parallax background is still working fine. Did notice a bug with the first puzzle where Active By Default is checked for a moving platform but the platform does not move, reached out to person who implemented it and will create a task later if necessary.